Cambio estructural a la organización del código:
- Se separaron los comandos del bot en la carpeta
  `app/cogs` para facilitar su mantenimiento y lectura.
- Se creó un nuevo archivo `metrics.py` para
  manejar los comandos relacionados con las métricas.
- Se creó un nuevo archivo `events.py` para
  manejar los eventos del bot.
- Se creó un nuevo archivo `general.py` para
  manejar los comandos generales del bot.
- Se modificó el archivo `app/main.py` para cargar
  los cogs de manera dinámica.

Cambio para eliminar código muerto y optimizar el bot:

- Se eliminó el uso de la función `setup_bot` ya
  que no era necesaria para la compatibilidad con
  singleton.
- Se eliminó la declaración de `voice_content` ya que
  no estaba en uso.
- Se utilizan los decoradores `@commands.Cog.listener()`
  y `@discord.slash_command()` para modernizar el
  modo de uso del bot y usar slash commands.

Cambios menores:

- Se reemplaza el prefijo por defecto del bot `!` a `/`
  por compatibilidad con los slash commands.
