+++ app/cogs/general.py
@@ -1,0 +1,15 @@
+import discord
+from discord.ext import commands
+
+class General(commands.Cog, name="GeneralCog"):
+    def __init__(self, bot):
+        self.bot = bot
+
+    @discord.slash_command(name="help",aliases=["h","ayuda"])
+    async def helper(self, ctx):
+        embed_help=discord.Embed(title="**Helper**", color=0xff0000)
+        embed_help.add_field(name="**Test**", value="**Prueba ok**", inline=True)
+        await typing(ctx,embed=embed_help,time=1)
+
+def setup(bot):
+    bot.add_cog(General(bot))
